pyproject.toml
[tool.poetry]
name = "multilayer_caching_memcached"
version = "0.1.3"
description = ""
authors = ["Amirhossein Faghihifar <a.faghihifar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
Flask-Cors = "^3.0.10"
Flask-RESTful = "^0.3.9"
requests = "^2.27.1"
gunicorn = "^20.0.4"
loguru = "^0.6.0"
elasticsearch = "^8.3.3"
PyYAML = "^6.0"
pymemcache = "^3.5.2"
SQLAlchemy = "^1.4.40"
psycopg2-binary = "^2.9.3"
pandas = "^1.4.3"


[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"